requirements.txt
fastapi
uvicorn[standard]
sqlmodel
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
python-multipart
pytest
httpx
pydantic-settings
